LayoutTests/http/tests/referrer-policy/no-referrer-when-downgrade/cross-origin-http-UpgradeMixedContent.https.html

<!DOCTYPE html><!-- webkit-test-runner [ UpgradeMixedContentEnabled=true ] -->
<html>
<head>
<meta name='referrer' content='no-referrer-when-downgrade'>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests the behavior of no-referrer-when-downgrade referrer policy when cross origin from HTTPS to HTTP.");
jsTestIsAsync = true;

if (window.internals) {
    internals.settings.setAllowDisplayOfInsecureContent(true);
    internals.setTrackingPreventionEnabled(false);
}

window.onmessage = function(event) {
    referrer = event.data.referrer;
    // Should be the empty string
    shouldBeEqualToString("referrer", "");
    if (window.internals)
        internals.setTrackingPreventionEnabled(true);
    finishJSTest();
}

setTimeout(() => {
    if (window.internals)
        internals.setTrackingPreventionEnabled(true);
    testPassed("Timeout reached, insecure iframe not loaded");
    finishJSTest();
}, 500);

if (window.testRunner) {
    testRunner.setStatisticsShouldDowngradeReferrer(false, function () {
        let iframeElement = document.createElement("iframe");
        iframeElement.src = "http://localhost:8000/referrer-policy/resources/document.html";
        document.body.appendChild(iframeElement);
    });
}
</script>
</body>
</html>
